Refactor and add README
